telegram_bot/handlers/calendar_handler.py
"""
Calendar-related handlers.
"""
import logging
from telegram import Update
from telegram.ext import ContextTypes

logger = logging.getLogger(__name__)


async def calendar_command(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
    """Handle /calendar command."""
    await update.message.reply_text(
        "Calendar integration coming soon!\n\n"
        "I'll help you:\n"
        "- View your calendar\n"
        "- Schedule tasks\n"
        "- Detect conflicts\n"
        "- Sync with Google Calendar"
    )

