requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Telegram Bot
python-telegram-bot==20.7
python-telegram-bot[job-queue]==20.7

# AI & LLM
langchain==0.1.0
langchain-google-genai==0.0.6
langchain-openai==0.0.2
langchain-community==0.0.10
llama-index==0.9.30
llama-index-vector-stores-postgres==0.1.4
llama-index-embeddings-openai==0.1.5

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Vector Database (Supabase pgvector)
pgvector==0.2.4

# Google Calendar API
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# Scheduling
apscheduler==3.10.4

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiohttp==3.9.1
pytz==2023.3
python-dateutil==2.8.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

